refactor: complete Frontend Modernization and DevOps configurations
