+++ db/install.xml
@@ -18,7 +18,7 @@
         <FIELD NAME="session_id" TYPE="char" LENGTH="255" NOTNULL="false" SEQUENCE="false" COMMENT="Session identifier linking related evidence"/>
         <FIELD NAME="ip_address" TYPE="char" LENGTH="45" NOTNULL="false" SEQUENCE="false" COMMENT="IP address for security tracking"/>
         <FIELD NAME="user_agent" TYPE="text" NOTNULL="false" SEQUENCE="false" COMMENT="Browser and device information"/>
-        <FIELD NAME="timestamp" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Unix timestamp when violation was detected"/>
+        <FIELD NAME="timecreated" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Unix timecreated when violation was detected"/>
         <FIELD NAME="severity" TYPE="char" LENGTH="20" NOTNULL="false" SEQUENCE="false" COMMENT="Severity level: low, medium, high, critical"/>
         <FIELD NAME="status" TYPE="char" LENGTH="20" NOTNULL="true" DEFAULT="active" SEQUENCE="false" COMMENT="Status: active, reviewed, dismissed, escalated"/>
         <FIELD NAME="reviewed_by" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="Teacher/admin who reviewed"/>
@@ -38,10 +38,10 @@
       <INDEXES>
         <INDEX NAME="course_user_idx" UNIQUE="false" FIELDS="courseid,userid" COMMENT="Find violations by course and user"/>
         <INDEX NAME="course_status_idx" UNIQUE="false" FIELDS="courseid,status" COMMENT="Find violations by course and status"/>
-        <INDEX NAME="severity_time_idx" UNIQUE="false" FIELDS="severity,timestamp" COMMENT="Sort by severity and time"/>
-        <INDEX NAME="user_time_idx" UNIQUE="false" FIELDS="userid,timestamp" COMMENT="User violation timeline"/>
+        <INDEX NAME="severity_time_idx" UNIQUE="false" FIELDS="severity,timecreated" COMMENT="Sort by severity and time"/>
+        <INDEX NAME="user_time_idx" UNIQUE="false" FIELDS="userid,timecreated" COMMENT="User violation timeline"/>
         <INDEX NAME="session_idx" UNIQUE="false" FIELDS="session_id" COMMENT="Group evidence by session"/>
-        <INDEX NAME="violation_type_idx" UNIQUE="false" FIELDS="violation_type,timestamp" COMMENT="Analyze violation patterns"/>
+        <INDEX NAME="violation_type_idx" UNIQUE="false" FIELDS="violation_type,timecreated" COMMENT="Analyze violation patterns"/>
       </INDEXES>
     </TABLE>
     
@@ -52,8 +52,8 @@
         <FIELD NAME="userid" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="User-specific settings (NULL for course defaults)"/>
         <FIELD NAME="setting_name" TYPE="char" LENGTH="100" NOTNULL="true" SEQUENCE="false" COMMENT="Configuration setting name"/>
         <FIELD NAME="setting_value" TYPE="text" NOTNULL="false" SEQUENCE="false" COMMENT="JSON or text configuration value"/>
-        <FIELD NAME="created_at" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Timestamp when setting was created"/>
-        <FIELD NAME="updated_at" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="Timestamp when setting was last modified"/>
+        <FIELD NAME="timecreated" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Timestamp when setting was created"/>
+        <FIELD NAME="timemodified" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="Timestamp when setting was last modified"/>
       </FIELDS>
       <KEYS>
         <KEY NAME="primary" TYPE="primary" FIELDS="id"/>
@@ -73,15 +73,15 @@
         <FIELD NAME="courseid" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Course where monitoring occurred"/>
         <FIELD NAME="userid" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="User being monitored"/>
         <FIELD NAME="session_id" TYPE="char" LENGTH="255" NOTNULL="true" SEQUENCE="false" COMMENT="Unique session identifier"/>
-        <FIELD NAME="starttime" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Session start timestamp"/>
-        <FIELD NAME="endtime" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="Session end timestamp"/>
+        <FIELD NAME="timestarted" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Session start timecreated"/>
+        <FIELD NAME="timeended" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="Session end timecreated"/>
         <FIELD NAME="duration" TYPE="int" LENGTH="10" NOTNULL="false" SEQUENCE="false" COMMENT="Total session duration in seconds"/>
         <FIELD NAME="violationcount" TYPE="int" LENGTH="10" NOTNULL="true" DEFAULT="0" SEQUENCE="false" COMMENT="Total violations detected"/>
         <FIELD NAME="evidencecount" TYPE="int" LENGTH="10" NOTNULL="true" DEFAULT="0" SEQUENCE="false" COMMENT="Total evidence files captured"/>
         <FIELD NAME="status" TYPE="char" LENGTH="30" NOTNULL="true" DEFAULT="active" SEQUENCE="false" COMMENT="Session status: active, completed, terminated, suspended"/>
         <FIELD NAME="systeminfo" TYPE="text" NOTNULL="false" SEQUENCE="false" COMMENT="JSON: browser, device, IP, etc."/>
         <FIELD NAME="finaldata" TYPE="text" NOTNULL="false" SEQUENCE="false" COMMENT="JSON: final metrics, suspicion levels, analysis"/>
-        <FIELD NAME="last_activity" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Last activity timestamp for cleanup"/>
+        <FIELD NAME="last_activity" TYPE="int" LENGTH="10" NOTNULL="true" SEQUENCE="false" COMMENT="Last activity timecreated for cleanup"/>
       </FIELDS>
       <KEYS>
         <KEY NAME="primary" TYPE="primary" FIELDS="id"/>
@@ -92,7 +92,7 @@
         <INDEX NAME="session_id_idx" UNIQUE="true" FIELDS="session_id" COMMENT="Unique session lookup"/>
         <INDEX NAME="course_user_idx" UNIQUE="false" FIELDS="courseid,userid" COMMENT="Find user sessions in course"/>
         <INDEX NAME="course_status_idx" UNIQUE="false" FIELDS="courseid,status" COMMENT="Find sessions by status"/>
-        <INDEX NAME="user_time_idx" UNIQUE="false" FIELDS="userid,starttime" COMMENT="User session timeline"/>
+        <INDEX NAME="user_time_idx" UNIQUE="false" FIELDS="userid,timestarted" COMMENT="User session timeline"/>
         <INDEX NAME="last_activity_idx" UNIQUE="false" FIELDS="last_activity" COMMENT="Session cleanup by activity"/>
         <INDEX NAME="violation_count_idx" UNIQUE="false" FIELDS="violationcount" COMMENT="Sessions by violation count"/>
       </INDEXES>
